+++ HISTORY.md
@@ -47,6 +47,7 @@
 * GUI: deprecated storing projects in localStorage
 * GUI: reenabled publish / unpublish buttons in the project dialog
 * GUI: spread project dialog action buttons over 2 rows
+* GUI: changed project dialog's initial & miminum extent
 
 ### 2019-02-26
 * Symbols: new "globe" symbol